+++ src/routes/api.php
@@ -1,9 +1,9 @@
 <?php
 
+use App\Domain\Auth\Http\Controllers\AuthController;
+use App\Domain\User\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Api\UserController;
-use App\Http\Controllers\Api\AuthController;
 
 Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);